requirements.txt
streamlit
Pillow
tensorflow
beautifulsoup4